ZOOKEEPER-2924: Refactor tests of LoadFromLogTest.java
Tests which don't need ZK test server to run have been extracted to
LoadFromLogNoServerTest.java.
Rest of them have been refactored to do init/shutdown in JUnit
Before/After methods.

Author: Andor Molnar <andor@cloudera.com>

Reviewers: phunt@apache.org

Closes #409 from anmolnar/ZOOKEEPER-2924
